# backend/src/main/java/io/swagger/api/BugApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.13).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.annotations.*;
import io.swagger.model.Bug;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-04-06T17:23:51.744Z")

@Api(value = "bug", description = "the bug API")
@RequestMapping(value = "")
public interface BugApi {

    @ApiOperation(value = "", nickname = "bugGet", notes = "Fetch all bugs ```Example: http://192.168.99.101:4041/bug```", response = Bug.class, responseContainer = "List", tags = {"Bug",})
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "Successfully fetched all bugs", response = Bug.class, responseContainer = "List")})
    @RequestMapping(value = "/bug",
            produces = {"application/json"},
            method = RequestMethod.GET)
    ResponseEntity<List<Bug>> bugGet();


    @ApiOperation(value = "", nickname = "bugNameGet", notes = "Fetch bug with passed name ```Example: http://192.168.99.101:4041/bug/tarantula```", response = Bug.class, tags = {"Bug",})
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "Successfully fetched a bug", response = Bug.class),
            @ApiResponse(code = 404, message = "The bug was not found")})
    @RequestMapping(value = "/bug/{name}",
            produces = {"application/json", "text/html"},
            method = RequestMethod.GET)
    ResponseEntity<Bug> bugNameGet(@ApiParam(value = "Name of the bug", required = true) @PathVariable("name") String name);

}